requirements.txt
accelerate==0.19.0
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
filelock==3.12.0
fsspec==2023.5.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
peft==0.3.0
psutil==5.9.5
python-dotenv==1.0.0
python-telegram-bot==20.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sniffio==1.3.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
urllib3==2.0.2
